requirements.txt
django=2
RPi.GPIO
